Revert "Add Cloud Events support for #55 (#56) (#64)"
This reverts commit 8f3fe352a5aa909251fb4b8bb6b417b2e96a4beb.
