--- examples/cloud_run_cloudevents/requirements.txt
@@ -1,1 +1,0 @@
-# Optionally include additional dependencies here
