--- examples/cloud_run_event/README.md
@@ -1,3 +1,0 @@
-# Google Cloud Functions Events Example
-This example demonstrates how to write an event function. Note that you can also use [CloudEvents](https://github.com/cloudevents/sdk-python)
-([example](../cloud_run_cloudevents)), which is a different construct.
