+++ src/functions_framework/__init__.py
@@ -12,19 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import enum
+import functools
 import importlib.util
-import io
-import json
 import os.path
 import pathlib
 import sys
 import types
 
-import cloudevents.sdk
-import cloudevents.sdk.event
-import cloudevents.sdk.event.v1
-import cloudevents.sdk.marshaller
 import flask
 import werkzeug
 
@@ -40,12 +34,6 @@ from google.cloud.functions.context import Context
 DEFAULT_SOURCE = os.path.realpath("./main.py")
 DEFAULT_SIGNATURE_TYPE = "http"
 MAX_CONTENT_LENGTH = 10 * 1024 * 1024
-
-
-class _EventType(enum.Enum):
-    LEGACY = 1
-    CLOUDEVENT_BINARY = 2
-    CLOUDEVENT_STRUCTURED = 3
 
 
 class _Event(object):
@@ -80,83 +68,38 @@ def _http_view_func_wrapper(function, request):
     return view_func
 
 
-def _get_cloudevent_version():
-    return cloudevents.sdk.event.v1.Event()
-
-
-def _run_legacy_event(function, request):
-    event_data = request.get_json()
-    if not event_data:
-        flask.abort(400)
-    event_object = _Event(**event_data)
-    data = event_object.data
-    context = Context(**event_object.context)
-    function(data, context)
-
-
-def _run_binary_cloudevent(function, request, cloudevent_def):
-    data = io.BytesIO(request.get_data())
-    http_marshaller = cloudevents.sdk.marshaller.NewDefaultHTTPMarshaller()
-    event = http_marshaller.FromRequest(
-        cloudevent_def, request.headers, data, json.load
-    )
-
-    function(event)
-
-
-def _run_structured_cloudevent(function, request, cloudevent_def):
-    data = io.StringIO(request.get_data(as_text=True))
-    m = cloudevents.sdk.marshaller.NewDefaultHTTPMarshaller()
-    event = m.FromRequest(cloudevent_def, request.headers, data, json.loads)
-    function(event)
-
-
-def _get_event_type(request):
-    if (
+def _is_binary_cloud_event(request):
+    return (
         request.headers.get("ce-type")
         and request.headers.get("ce-specversion")
         and request.headers.get("ce-source")
         and request.headers.get("ce-id")
-    ):
-        return _EventType.CLOUDEVENT_BINARY
-    elif request.headers.get("Content-Type") == "application/cloudevents+json":
-        return _EventType.CLOUDEVENT_STRUCTURED
-    else:
-        return _EventType.LEGACY
+    )
 
 
 def _event_view_func_wrapper(function, request):
     def view_func(path):
-        if _get_event_type(request) == _EventType.LEGACY:
-            _run_legacy_event(function, request)
-        else:
-            # here for defensive backwards compatibility in case we make a mistake in rollout.
-            flask.abort(
-                400,
-                description="The FUNCTION_SIGNATURE_TYPE for this function is set to event "
-                "but no Google Cloud Functions Event was given. If you are using CloudEvents set "
-                "FUNCTION_SIGNATURE_TYPE=cloudevent",
+        if _is_binary_cloud_event(request):
+            # Support CloudEvents in binary content mode, with data being the
+            # whole request body and context attributes retrieved from request
+            # headers.
+            data = request.get_data()
+            context = Context(
+                eventId=request.headers.get("ce-eventId"),
+                timestamp=request.headers.get("ce-timestamp"),
+                eventType=request.headers.get("ce-eventType"),
+                resource=request.headers.get("ce-resource"),
             )
-
-        return "OK"
-
-    return view_func
-
-
-def _cloudevent_view_func_wrapper(function, request):
-    def view_func(path):
-        cloudevent_def = _get_cloudevent_version()
-        event_type = _get_event_type(request)
-        if event_type == _EventType.CLOUDEVENT_STRUCTURED:
-            _run_structured_cloudevent(function, request, cloudevent_def)
-        elif event_type == _EventType.CLOUDEVENT_BINARY:
-            _run_binary_cloudevent(function, request, cloudevent_def)
+            function(data, context)
         else:
-            flask.abort(
-                400,
-                description="Function was defined with FUNCTION_SIGNATURE_TYPE=cloudevent "
-                " but it did not receive a cloudevent as a request.",
-            )
+            # This is a regular CloudEvent
+            event_data = request.get_json()
+            if not event_data:
+                flask.abort(400)
+            event_object = _Event(**event_data)
+            data = event_object.data
+            context = Context(**event_object.context)
+            function(data, context)
 
         return "OK"
 
@@ -263,27 +206,19 @@ def create_app(target=None, source=None, signature_type=None):
         app.view_functions["run"] = _http_view_func_wrapper(function, flask.request)
         app.view_functions["error"] = lambda: flask.abort(404, description="Not Found")
         app.after_request(read_request)
-    elif signature_type == "event" or signature_type == "cloudevent":
+    elif signature_type == "event":
         app.url_map.add(
             werkzeug.routing.Rule(
-                "/", defaults={"path": ""}, endpoint=signature_type, methods=["POST"]
+                "/", defaults={"path": ""}, endpoint="run", methods=["POST"]
             )
         )
         app.url_map.add(
-            werkzeug.routing.Rule(
-                "/<path:path>", endpoint=signature_type, methods=["POST"]
-            )
+            werkzeug.routing.Rule("/<path:path>", endpoint="run", methods=["POST"])
         )
-
+        app.view_functions["run"] = _event_view_func_wrapper(function, flask.request)
         # Add a dummy endpoint for GET /
         app.url_map.add(werkzeug.routing.Rule("/", endpoint="get", methods=["GET"]))
         app.view_functions["get"] = lambda: ""
-
-        # Add the view functions
-        app.view_functions["event"] = _event_view_func_wrapper(function, flask.request)
-        app.view_functions["cloudevent"] = _cloudevent_view_func_wrapper(
-            function, flask.request
-        )
     else:
         raise FunctionsFrameworkException(
             "Invalid signature type: {signature_type}".format(
